# docs/swagger.yaml
definitions:
...
  github_com_evrone_go-clean-template_internal_controller_restapi_v1_request.ConfirmUpload:
    properties:
      video_id:
        type: string
    required:
    - video_id
    type: object
  github_com_evrone_go-clean-template_internal_controller_restapi_v1_request.CreateVideoUpload:
    properties:
      category:
        type: string
      description:
        maxLength: 1000
        type: string
      file_name:
        type: string
      title:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - file_name
    - title
    type: object
  github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.Error:
    properties:
      error:
        example: internal server error
        type: string
    type: object
...
  github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.SystemDashboardResponse:
    properties:
      active_livestreams:
        example: 12
        type: integer
      active_workers:
        example: 4
        type: integer
      bandwidth_usage_gb:
        example: 128.4
        type: number
      total_videos:
        example: 1420
        type: integer
      total_viewers:
        example: 35480
        type: integer
    type: object
  github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.UploadUrlResponse:
    properties:
      raw_s3_key:
        type: string
      upload_url:
        type: string
      video_id:
        type: string
    type: object
  github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.VideoResponse:
    properties:
      category:
        example: Science
        type: string
      created_at:
        type: string
      description:
        example: Join Luna Chen as she explores space.
        type: string
      duration:
        example: "18:42"
        type: string
      hls_url:
        example: https://s3.pipevid.com/hls/space/master.m3u8
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      status:
        example: complete
        type: string
      thumbnail_url:
        example: https://s3.pipevid.com/thumbnails/space.jpg
        type: string
      title:
        example: Exploring the edge of the universe
        type: string
      updated_at:
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      views:
        example: 24812
        type: integer
      visibility:
        example: public
        type: string
    type: object
  github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.WorkerStatusResponse:
    properties:
      cpu_usage:
        example: 42.5
        type: number
      current_job:
        example: job_vid_9421
        type: string
      ram_usage:
        example: 68.1
        type: number
      status:
        example: processing
        type: string
      worker_id:
        example: worker-node-1
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /v1/admin/dashboard:
    get:
      consumes:
      - application/json
      description: Get system overview statistics (Total Videos, Active Streams, Viewers,
        Bandwidth)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.SystemDashboardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.Error'
      security:
      - BearerAuth: []
      summary: Get Admin Dashboard Metrics
      tags:
      - Admin
  /v1/admin/streams/{id}/ban:
    post:
      consumes:
      - application/json
      description: Instantly terminate and ban an active livestream
      parameters:
      - description: Livestream ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.Error'
      security:
      - BearerAuth: []
      summary: Ban active livestream
      tags:
      - Admin
  /v1/admin/videos/{id}/ban:
    post:
      consumes:
      - application/json
      description: Remove or set video visibility to private due to terms violation
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.Error'
      security:
      - BearerAuth: []
      summary: Ban video
      tags:
      - Admin
  /v1/admin/workers:
    get:
      consumes:
      - application/json
      description: Get real-time CPU, RAM, and current processing jobs of transcode
        worker nodes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.WorkerStatusResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.Error'
      security:
      - BearerAuth: []
      summary: Get Transcode Workers Status
      tags:
      - Admin
...
  /v1/studio/confirm-upload:
    post:
      consumes:
      - application/json
      description: Confirm that video file upload to S3 is complete and trigger transcode
        worker
      parameters:
      - description: Confirm Upload Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_request.ConfirmUpload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.VideoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.Error'
      security:
      - BearerAuth: []
      summary: Confirm raw video upload
      tags:
      - Studio
...
  /v1/studio/upload-url:
    post:
      consumes:
      - application/json
      description: Get a presigned S3 URL to upload raw video
      parameters:
      - description: Video Upload Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_request.CreateVideoUpload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.UploadUrlResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.Error'
      security:
      - BearerAuth: []
      summary: Request Presigned S3 Upload URL
      tags:
      - Studio
  /v1/studio/videos:
    get:
      consumes:
      - application/json
      description: Get videos uploaded by the authenticated streamer
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.VideoResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.Error'
      security:
      - BearerAuth: []
      summary: List studio videos
      tags:
      - Studio
  /v1/studio/videos/{id}/publish:
    post:
      consumes:
      - application/json
      description: Set video visibility to public
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.VideoResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.Error'
      security:
      - BearerAuth: []
      summary: Publish studio video
      tags:
      - Studio
  /v1/videos:
    get:
      consumes:
      - application/json
      description: Get a paginated list of public videos with optional category filter
      parameters:
      - description: Filter by category
        in: query
        name: category
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.VideoResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.Error'
      summary: List public videos
      tags:
      - Videos
  /v1/videos/{id}:
    get:
      consumes:
      - application/json
      description: Get video details by ID
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.VideoResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_restapi_v1_response.Error'
      summary: Get video by ID
      tags:
      - Videos
swagger: "2.0"